refactor code after the review
