+++ ethereum/statetransition/build.gradle
@@ -25,9 +25,6 @@ dependencies {
   implementation project(':storage')
   implementation project(':storage:api')
 
-  implementation testFixtures(project(':infrastructure:metrics'))
-
-  implementation 'io.prometheus:simpleclient'
   implementation 'org.apache.tuweni:tuweni-units'
   implementation 'org.hyperledger.besu.internal:metrics-core'
 
@@ -37,10 +34,8 @@ dependencies {
   testImplementation testFixtures(project(':infrastructure:async'))
   testImplementation testFixtures(project(':infrastructure:bls'))
   testImplementation testFixtures(project(':infrastructure:time'))
-  testImplementation testFixtures(project(':infrastructure:metrics'))
   testImplementation testFixtures(project(':infrastructure:kzg'))
   testImplementation testFixtures(project(':storage'))
-  testImplementation 'org.hyperledger.besu.internal:metrics-core'
 
   testFixturesApi testFixtures(project(':infrastructure:bls'))
   testFixturesApi testFixtures(project(':storage'))
